+++ types/block.go
@@ -658,7 +658,7 @@ type Commit struct {
 	// Any peer with a block can gossip signatures by index with a peer without
 	// recalculating the active ValidatorSet.
 	Height     int64       `json:"height"`
-	Round      int         `json:"round"`
+	Round      int32       `json:"round"`
 	BlockID    BlockID     `json:"block_id"`
 	Signatures []CommitSig `json:"signatures"`
 
@@ -670,7 +670,7 @@ type Commit struct {
 }
 
 // NewCommit returns a new Commit.
-func NewCommit(height int64, round int, blockID BlockID, commitSigs []CommitSig) *Commit {
+func NewCommit(height int64, round int32, blockID BlockID, commitSigs []CommitSig) *Commit {
 	return &Commit{
 		Height:     height,
 		Round:      round,
@@ -688,7 +688,7 @@ func CommitToVoteSet(chainID string, commit *Commit, vals *ValidatorSet) *VoteSe
 		if commitSig.Absent() {
 			continue // OK, some precommits can be missing.
 		}
-		added, err := voteSet.AddVote(commit.GetVote(idx))
+		added, err := voteSet.AddVote(commit.GetVote(int32(idx)))
 		if !added || err != nil {
 			panic(fmt.Sprintf("Failed to reconstruct LastCommit: %v", err))
 		}
@@ -699,7 +699,7 @@ func CommitToVoteSet(chainID string, commit *Commit, vals *ValidatorSet) *VoteSe
 // GetVote converts the CommitSig for the given valIdx to a Vote.
 // Returns nil if the precommit at valIdx is nil.
 // Panics if valIdx >= commit.Size().
-func (commit *Commit) GetVote(valIdx int) *Vote {
+func (commit *Commit) GetVote(valIdx int32) *Vote {
 	commitSig := commit.Signatures[valIdx]
 	return &Vote{
 		Type:             PrecommitType,
@@ -717,7 +717,7 @@ func (commit *Commit) GetVote(valIdx int) *Vote {
 // The only unique part of the SignBytes is the Timestamp - all other fields
 // signed over are otherwise the same for all validators.
 // Panics if valIdx >= commit.Size().
-func (commit *Commit) VoteSignBytes(chainID string, valIdx int) []byte {
+func (commit *Commit) VoteSignBytes(chainID string, valIdx int32) []byte {
 	return commit.GetVote(valIdx).SignBytes(chainID)
 }
 
@@ -735,7 +735,7 @@ func (commit *Commit) GetHeight() int64 {
 
 // GetRound returns height of the commit.
 // Implements VoteSetReader.
-func (commit *Commit) GetRound() int {
+func (commit *Commit) GetRound() int32 {
 	return commit.Round
 }
 
@@ -765,7 +765,7 @@ func (commit *Commit) BitArray() *bits.BitArray {
 // GetByIndex returns the vote corresponding to a given validator index.
 // Panics if `index >= commit.Size()`.
 // Implements VoteSetReader.
-func (commit *Commit) GetByIndex(valIdx int) *Vote {
+func (commit *Commit) GetByIndex(valIdx int32) *Vote {
 	return commit.GetVote(valIdx)
 }
 
@@ -892,7 +892,7 @@ func CommitFromProto(cp *tmproto.Commit) (*Commit, error) {
 	commit.Signatures = sigs
 
 	commit.Height = cp.Height
-	commit.Round = int(cp.Round)
+	commit.Round = cp.Round
 	commit.BlockID = *bi
 	commit.hash = cp.Hash
 	commit.bitArray = bitArray
